ADD: modificaciones menores para gestion de ventana

+++ esp/main/bme.c
@@ -342,8 +342,8 @@ int bme_temp_celsius(uint32_t temp_adc) {
 }
 
 int bme_pres_pa(uint32_t pres_adc) {
-    // Datasheet[23]
-    // https://www.bosch-sensortec.com/media/boschsensortec/downloads/datasheets/bst-bme688 var2 = (var2 / 4.0f) + (((float)dev->calib.par_p4) * 65536.0f);-ds000.pdf#page=23
+    // Datasheet[24]
+    // https://www.bosch-sensortec.com/media/boschsensortec/downloads/datasheets/bst-bme688-ds000.pdf#page=24 var2 = (var2 / 4.0f) + (((float)dev->calib.par_p4) * 65536.0f);-ds000.pdf#page=23
 
     // Se obtienen los parametros de calibracion
     uint8_t addr_par_p1_lsb = 0x8E, addr_par_p1_msb = 0x8F;
@@ -411,17 +411,17 @@ int bme_pres_pa(uint32_t pres_adc) {
     var1 = var1 >> 18;
     var1 = ((32768 + var1) * (int32_t)par_p1) >> 15;
     calc_pres = 1048576 - pres_adc;
-    calc_pres = (int)((calc_pres - (var2 >> 12)) * ((uint32_t)3125));
+    calc_pres = (int32_t)((calc_pres - (var2 >> 12)) * ((uint32_t)3125));
     if (calc_pres >= (1 << 30)) {
-        calc_pres = ((calc_pres / (uint32_t)var1) << 1);
+        calc_pres = ((calc_pres / var1) << 1);
     } else {
-        calc_pres = ((calc_pres << 1) / (uint32_t)var1);
+        calc_pres = ((calc_pres << 1) / var1);
     }
     var1 = ((int32_t)par_p9 * (int32_t)(((calc_pres >> 3) * (calc_pres >> 3)) >> 13)) >> 12;
     var2 = ((int32_t)(calc_pres >> 2) * (int32_t)par_p8) >> 13;
     var3 = ((int32_t)(calc_pres >> 8) * (int32_t)(calc_pres >> 8) * (int32_t)(calc_pres >> 8) * (int32_t)par_p10) >> 17;
     calc_pres = (int32_t)(calc_pres) + ((var1 + var2 + var3 + ((int32_t)par_p7 << 7)) >> 4);
-    return calc_pres;
+    return (int)calc_pres;
 }
 
 void bme_get_mode(void) {
